+++ app/src/main/java/the/head/that/feeds/Stats.kt
@@ -3,7 +3,7 @@ package the.head.that.feeds
 //Todo: Stats need to have separate data class + database/sharedPref access. We're just using LiveData to observe state so we can update UI.
 
 class Stats {
-    var friendlyAIIntegrity : Double = 0.0
+    var friendlyAIIntegrity : Int = 0
     var friendlyAIEvolutionLevel : Int = 0
     var aggression : Int = 0
     var empathy : Int = 0
@@ -11,13 +11,13 @@ class Stats {
     var fighters : Int = 0
     var civilians : Double = 0.0
 
-    var gridAIIntegrity : Double = 0.0
+    var gridAIIntegrity : Int = 0
     var gridAITrackingLevel : Int = 0
 
     var currentDay : Int = 0
 
     fun setDefaultStatValues() {
-        friendlyAIIntegrity = 100.0
+        friendlyAIIntegrity = 100
         friendlyAIEvolutionLevel = 0
         aggression = 20
         empathy = 20
@@ -25,7 +25,7 @@ class Stats {
         fighters = 100
         civilians = 42.0
 
-        gridAIIntegrity = 100.0
+        gridAIIntegrity = 100
         gridAITrackingLevel = 0
     }
 }
